packages/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
boto3==1.37.8
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.11
h11==0.14.0
httpx==0.28.1
idna==3.10
jose==1.0.0
python-jose==3.4.0
legacy-cgi==2.6.2; python_version >= "3.13"
litellm==1.67.6
mongomock-motor==0.0.35
motor==3.7.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pymongo==4.11.2
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
pytest-asyncio==0.25.3
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
stripe==11.6.0
typing_extensions==4.12.2
uvicorn==0.34.0
